installer: install wintun.dll to system32
The problem with this approach is the EvaluateWintun action is executed
*before* the wintun.dll is really installed - in the immediate MSI
phase. Therefore, it fails.

All immediate custom actions have to be based on a DLL stored in the
Binary table.

We should reintroduce installer.dll in the Binary table just to
implement the EvaluateWintun action - which voids the idea to have a
one-size-fits-all DLL.

Signed-off-by: Simon Rozman <simon@rozman.si>